requirements.txt
# Core ML dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# Data processing and analysis
pandas>=1.4.0
scikit-learn>=1.0.0

# Image processing
opencv-python>=4.6.0
pillow>=9.0.0
albumentations>=1.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress tracking and UI
tqdm>=4.64.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.9.0

# Model visualization and analysis
torchsummary>=1.5.1
torchview>=0.2.0
graphviz>=0.20.0

# Object detection (YOLO integration)
ultralytics>=8.0.0